requirements.txt
apache-airflow==2.8.2
apache-airflow-providers-apache-spark
pyspark==3.5.0